refactor(mobile): rename use-kiloclaw to use-kiloclaw-queries
