+++ apps/mobile/src/app/(app)/chat/[instance-id].tsx
@@ -3,7 +3,7 @@ import { View } from 'react-native';
 
 import { KiloClawChat } from '@/components/kiloclaw/chat';
 import { useInstanceContext } from '@/lib/hooks/use-instance-context';
-import { useKiloClawStatus } from '@/lib/hooks/use-kiloclaw';
+import { useKiloClawStatus } from '@/lib/hooks/use-kiloclaw-queries';
 
 export default function ChatScreen() {
   const { 'instance-id': instanceId } = useLocalSearchParams<{ 'instance-id': string }>();
